+++ lib/rig_api/router.ex
@@ -1,7 +1,7 @@
 defmodule RigApi.Router do
   use RigApi, :router
 
-  pipeline :api do
+  pipeline :body_parser do
     plug(Plug.Parsers,
       parsers: [:urlencoded, :multipart, :json],
       # return "415 Unsupported Media Type" if not handled by any parser
@@ -10,32 +10,9 @@ defmodule RigApi.Router do
     )
   end
 
-  scope "/v1/messages", RigApi do
-    post("/", MessageController, :publish)
-  end
-
-  scope "/v1", RigApi do
-    pipe_through(:api)
-
-    resources("/responses", ResponsesController, only: [:create])
-
-    scope "/session-blacklist" do
-      post("/", SessionBlacklistController, :blacklist_session)
-      get("/:session_id", SessionBlacklistController, :check_status)
-    end
-
-    scope "/apis" do
-      get("/", ApisController, :list_apis)
-      post("/", ApisController, :add_api)
-      get("/:id", ApisController, :get_api_detail)
-      put("/:id", ApisController, :update_api)
-      delete("/:id", ApisController, :deactivate_api)
-    end
-  end
-
   scope "/health", RigApi do
-    pipe_through(:api)
-    get("/", HealthController, :check_health)
+    pipe_through(:body_parser)
+    get("/", Health, :check_health)
   end
 
   scope "/swagger-ui" do
@@ -43,6 +20,59 @@ defmodule RigApi.Router do
       otp_app: :rig,
       swagger_file: "rig_api_swagger.json"
     )
+  end
+
+  # Deprecated in 2.3, to be removed with 3.0:
+  scope "/v1", RigApi.V1 do
+    scope "/apis" do
+      pipe_through(:body_parser)
+      get("/", APIs, :list_apis)
+      post("/", APIs, :add_api)
+      get("/:id", APIs, :get_api_detail)
+      put("/:id", APIs, :update_api)
+      delete("/:id", APIs, :deactivate_api)
+    end
+
+    scope "/messages" do
+      post("/", Messages, :publish)
+    end
+
+    scope "/responses" do
+      pipe_through(:body_parser)
+      resources("/", Responses, only: [:create])
+    end
+
+    scope "/session-blacklist" do
+      pipe_through(:body_parser)
+      post("/", SessionBlacklist, :blacklist_session)
+      get("/:session_id", SessionBlacklist, :check_status)
+    end
+  end
+
+  scope "/v2", RigApi.V2 do
+    scope "/apis" do
+      pipe_through(:body_parser)
+      get("/", APIs, :list_apis)
+      post("/", APIs, :add_api)
+      get("/:id", APIs, :get_api_detail)
+      put("/:id", APIs, :update_api)
+      delete("/:id", APIs, :deactivate_api)
+    end
+
+    scope "/messages" do
+      post("/", Messages, :publish)
+    end
+
+    scope "/responses" do
+      pipe_through(:body_parser)
+      resources("/", Responses, only: [:create])
+    end
+
+    scope "/session-blacklist" do
+      pipe_through(:body_parser)
+      post("/", SessionBlacklist, :blacklist_session)
+      get("/:session_id", SessionBlacklist, :check_status)
+    end
   end
 
   def swagger_info do